Incomplete proposal changes for reducing duplicate code on Rule Engine children
