+++ etor/src/main/java/gov/hhs/cdc/trustedintermediary/etor/ruleengine/validation/ValidationRuleEngine.java
@@ -1,28 +1,22 @@
 package gov.hhs.cdc.trustedintermediary.etor.ruleengine.validation;
 
-import gov.hhs.cdc.trustedintermediary.etor.ruleengine.FhirResource;
-import gov.hhs.cdc.trustedintermediary.etor.ruleengine.RuleEngine;
-import gov.hhs.cdc.trustedintermediary.etor.ruleengine.RuleLoader;
-import gov.hhs.cdc.trustedintermediary.etor.ruleengine.RuleLoaderException;
-import gov.hhs.cdc.trustedintermediary.wrappers.Logger;
-import gov.hhs.cdc.trustedintermediary.wrappers.formatter.TypeReference;
-import java.io.InputStream;
+import gov.hhs.cdc.trustedintermediary.etor.ruleengine.RuleEngineImplementation;
 import java.util.ArrayList;
 import java.util.List;
-import javax.inject.Inject;
 
 /** Implements the RuleEngine interface. It represents a rule engine for validations. */
-public class ValidationRuleEngine implements RuleEngine {
-    private String ruleDefinitionsFileName;
+public class ValidationRuleEngine extends RuleEngineImplementation {
     final List<ValidationRule> rules = new ArrayList<>();
 
     private static final ValidationRuleEngine INSTANCE = new ValidationRuleEngine();
 
-    @Inject Logger logger;
-    @Inject RuleLoader ruleLoader;
-
     public static ValidationRuleEngine getInstance(String ruleDefinitionsFileName) {
         INSTANCE.ruleDefinitionsFileName = ruleDefinitionsFileName;
+        return INSTANCE;
+    }
+
+    //    @Override
+    public ValidationRuleEngine getInstance() {
         return INSTANCE;
     }
 
@@ -31,47 +25,5 @@ public class ValidationRuleEngine implements RuleEngine {
     @Override
     public void unloadRules() {
         rules.clear();
-    }
-
-    @Override
-    public void ensureRulesLoaded() throws RuleLoaderException {
-        // Double-checked locking - needed to protect from excessive sync locks
-        if (rules.isEmpty()) {
-            synchronized (this) {
-                if (rules.isEmpty()) {
-                    InputStream resourceStream =
-                            getClass()
-                                    .getClassLoader()
-                                    .getResourceAsStream(ruleDefinitionsFileName);
-                    assert resourceStream != null;
-                    List<ValidationRule> parsedRules =
-                            ruleLoader.loadRules(resourceStream, new TypeReference<>() {});
-                    this.rules.addAll(parsedRules);
-                }
-            }
-        }
-    }
-
-    @Override
-    public void runRules(FhirResource<?> resource) {
-        try {
-            ensureRulesLoaded();
-        } catch (RuleLoaderException e) {
-            logger.logError("Failed to load rules definitions", e);
-            return;
-        }
-        for (ValidationRule rule : rules) {
-            if (rule.shouldRun(resource)) {
-                rule.runRule(resource);
-            }
-        }
-    }
-
-    @Override
-    public ValidationRule getRuleByName(String ruleName) {
-        return rules.stream()
-                .filter(rule -> rule.getName().equals(ruleName))
-                .findFirst()
-                .orElse(null);
     }
 }
